# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChangeEmailRequest" : {
        "properties" : {
          "newEmail" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "newEmail" ],
        "type" : "object"
      },
...
      "CodeSentDto" : {
        "description" : "The ack body of a 'code sent' endpoint (verification request, password-reset request, contact-change request).",
        "properties" : {
          "resendAvailableAfterSeconds" : {
            "description" : "Seconds the client should wait before asking for the same code again — the server-enforced resend cooldown (0 = no cooldown configured).",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ConfirmChangeRequest" : {
        "description" : "Body for POST /account/email-change/confirm and POST /account/phone-change/confirm.",
        "properties" : {
          "code" : {
            "description" : "The 6-digit code sent to the account's other contact channel.",
            "maxLength" : 16,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "code" ],
        "type" : "object"
      },
...
      "DataExportResponse" : {
        "properties" : {
          "profile" : {
            "$ref" : "#/components/schemas/ExportedProfile"
          },
          "shelters" : {
            "items" : {
              "$ref" : "#/components/schemas/ExportedShelter"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "MeResponse" : {
        "description" : "The authenticated user's real profile + REAL verification claims — the response of GET /account/me and PUT /account/profile.",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "isAdmin" : {
            "description" : "Always present: the frontend gates the admin route and nav item on it. Reflects the freshly loaded user's KIND (kind is the truth) — true only for the ADMIN-kind account, never derived from any token claim.",
            "type" : "boolean"
          },
          "levels" : {
            "description" : "Exactly the levels that are actually verified for this account (non-revoked claims), in enum order for a stable body — the frontend's single source of truth for its verification labels.",
            "items" : {
              "description" : "Exactly the levels that are actually verified for this account (non-revoked claims), in enum order for a stable body — the frontend's single source of truth for its verification labels.",
              "enum" : [ "EMAIL", "PHONE", "SMART_ID" ],
              "type" : "string"
            },
            "type" : "array"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The `Authorization: Bearer <accessToken>` header read by the JWT filter. A valid token of a SUSPENDED account authenticates nothing (fresh per-request DB lookup — the DB is the truth, never a token claim).",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "aleks.bratsun@reaktiiv.com",
      "name" : "Aleks Bratsun (project owner)"
    },
    "description" : "The machine-readable contract of the OpenShelter backend. Companion to the human contracts: context-and-tasks/05-shelter-api.puml (the read contract), the README 'API' section and docs/whitepaper.md. Served at /v3/api-docs in dev/test only (SPRINGDOC_ENABLED); the committed snapshot is docs/api/openapi.json. Every non-2xx response carries the uniform ErrorResponse body.",
    "license" : {
      "name" : "MIT"
    },
    "title" : "OpenShelter API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/account" : {
      "delete" : {
        "description" : "The account erasure: the declared private homes are purged, the public community rows are orphaned (map data outlives accounts), and the DB cascades credentials, claims, pending changes, tokens and reports. The verified-user gate matches the submission gates (403 without a claim); a repeat call is an idempotent no-op — the JWT is valid until its expiry, but the account is already gone. The environment-provisioned administrator (kind ADMIN) is refused with 403 naming the environment provisioning — de-provisioning is an operator action on the env vars, not an in-app deletion.",
        "operationId" : "deleteAccount",
        "responses" : {
          "204" : {
            "description" : "The account is erased (an idempotent no-op on a repeat call)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "description" : "No verification claim — or the environment-provisioned administrator account (refused; the message names the env provisioning)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Erase the account",
        "tags" : [ "Account & verification" ]
      }
    },
    "/account/email-change/confirm" : {
      "post" : {
        "description" : "Completes the change once the code is verified; a code failure is a 400.",
        "operationId" : "confirmEmailChange",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "E-mail changed"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Confirm the e-mail change",
        "tags" : [ "Account & verification" ]
      }
    },
    "/account/email-change/request" : {
      "post" : {
        "description" : "SMS code to the current phone (an email thief alone cannot change the email). 202 + the resend-cooldown ack.",
        "operationId" : "requestEmailChange",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeEmailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CodeSentDto"
                }
              }
            },
            "description" : "Code sent — the ack carries the resend cooldown in seconds"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Request an e-mail change",
        "tags" : [ "Account & verification" ]
      }
    },
    "/account/export" : {
      "get" : {
        "description" : "The caller's own data (profile + every author-scoped shelter row) as one JSON document. Same auth rule as /me (valid JWT, user from the token) and, like /me, no rate bucket (cheap read). The frontend turns the body into a downloadable file.",
        "operationId" : "dataExport",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DataExportResponse"
                }
              }
            },
            "description" : "The data export document"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "The caller's own data export",
        "tags" : [ "Account & verification" ]
      }
    },
    "/account/me" : {
      "get" : {
        "description" : "The user's real profile + REAL verification claims — the frontend's single source of truth for name/email/phone and verification labels. No rate bucket (cheap read).",
        "operationId" : "me",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "The profile"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "The caller's profile + verified claims",
        "tags" : [ "Account & verification" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
}